Corrigé quelques données des tests WIT suite au changement de référentiel des rues (IFOII -> FiDoR/RefInf)
git-svn-id: https://svn.etat-de-vaud.ch/SVN_SOCLE/registre/unireg/trunk@118477 cb3efa0a-b8b0-9f43-a4c8-343d8b7e4362

+++ unireg/web-it/src/test/java/ch/vd/uniregctb/webservice/party3/PartyWebServiceTest.java
@@ -312,7 +312,7 @@ public class PartyWebServiceTest extends AbstractPartyWebServiceTest {
 
 		final AddressInformation represInfo2 = repres2.getAddressInformation();
 		assertNotNull(represInfo2);
-		assertEquals(new Integer(32296), represInfo2.getStreetId());
+		assertEquals(new Integer(1131419), represInfo2.getStreetId());
 		assertEquals("1", represInfo2.getHouseNumber());
 		assertEquals(Integer.valueOf(528), represInfo2.getSwissZipCodeId());
 		assertEquals("Avenue du Funiculaire", represInfo2.getStreet());
